# docs/services/subscriptions/api/openApi.yaml
openapi: 3.1.0
info:
  title: RollFinders Subscription Service API
  version: 0.1.0
  description: >
    Contract for subscription products, product features, plans, owner
    subscriptions, plan changes, billing events, and entitlement decisions.
servers:
  - url: /v1
security:
  - gatewayAuthorisation: []
paths:
...
  /entitlements/check:
    post:
      summary: Check whether an owner has a subscription-controlled feature
      x-permission: subscription.entitlement.read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntitlementCheckRequest"
      responses:
        "200":
          description: Entitlement decision.
components:
  securitySchemes:
    gatewayAuthorisation:
      type: http
      scheme: bearer
      description: API Gateway forwards only after Authorisation Service permission checks.
...
  schemas:
...
    EntitlementCheckRequest:
      type: object
      required: [owner_type, owner_id, feature_key]
      properties:
        owner_type:
          type: string
        owner_id:
          type: string
        feature_key:
          type: string
        subject_id:
          type: string
        permission:
          type: string
        application_id:
          type: string
        organisation_id:
          type: string
        resource_id:
          type: string
        route:
          type: string
        http_method:
          type: string